--- src/components/Login/Login.scss
@@ -1,68 +1,0 @@
-.login-page {
-  margin: 0 auto;
-  padding: 0 1rem;
-
-  @media (min-width: 640px) {
-    max-width: 28rem;
-  }
-
-  header {
-    h1 {
-      margin-top: 3rem;
-      margin-bottom: 2rem;
-      text-align: center;
-      font-size: 1.25rem;
-      font-weight: 700;
-    }
-  }
-
-  form {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-
-    button {
-      width: 100%;
-    }
-
-    .error-alert {
-      text-align: center;
-      font-size: 0.875rem;
-      font-weight: 500;
-      color: #dc2626;
-    }
-  }
-
-  .gray-label {
-    margin: 2rem 0;
-    display: block;
-    text-align: center;
-    font-size: 0.875rem;
-    color: #7f7f7f;
-  }
-
-  .social-login {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-
-    h2 {
-      text-align: center;
-      font-weight: 500;
-      color: #6c6c6c;
-    }
-
-    .button-group {
-      margin-top: 1rem;
-      display: flex;
-      gap: 1rem;
-    }
-  }
-
-  .bottom {
-    margin-top: 2rem;
-    display: flex;
-    justify-content: center;
-    gap: 2rem;
-  }
-}
